Fixes #596 - pre-creates .rgw.buckets.extra for RADOSgw
RADOSgw uses .rgw.buckets.extra to store metadata when accepting
multipart file uploads. If this pool is not already created, it tries to
create it with crush_ruleset 0, size 3, min_size 2, and so the PGs will
never go active because crush_ruleset 0 is not one that exists in our
setup. This adds that pool to the list of pools to be pre-created.