Add back just the pre-built VM production part of the abandoned PR
https://github.com/bloomberg/chef-bcpc/pull/143

This is an entirely optional and separate addition which provides a
way to pre-build the chef-bcpc bootstrap node as an exported virtual
machine (using Packer) with all expected network configuration already
performed. It is much faster than building from scratch.

If the resulting .ova file is available, vbox_create.sh can
use it instead of creating that VM from scratch (if you define
ARCHIVED_BOOTSTRAP to point to the file, by default it's
undefined). After importing from the .ova, vbox_create.sh still runs its
NIC modifications as if it had just build the VM itself, this ensures
the VM is precisely as we expect in the non-vagrant path.

If you do not go into the new directory 'images' and run packer in
there to build, this PR makes absolutely zero changes to any of the
build procedures.

This version of the image production system has been updated for
Ubuntu 14.04.2 but could be used for Ubuntu 12.04 still (see the
README.md in /images).