Use pre-packed image if available.
Use an exported .ova file for the bootstrap node if it exists in this
file hierarchy. Since we now have the packer code on master, we know
where the image will be if you opted to go build it. If you don't want
to use it, just don't run that build step.