Intentionally fail Chef run to avoid cutting off management interface
If interface-parent for floating or storage is set to the management
interface, don't overwrite the network configuration for the management
interface. (This effectively means that the storage interface must be
on its own and management/floating must share a physical interface
because of MTU 9000 on the storage interface, unless you also want
to run MTU 9000 on the management interface).