Install/upgrade xtrabackup before MySQL
This change fixes a situation that can be encountered during MySQL
upgrade when Percona requires a new minimum version of xtrabackup (for
Galera synchronization). This specifically fixes situations that put
errors like the following in the MySQL error log on MySQL start:

"WSREP_SST: [ERROR] FATAL: The innobackupex version is 2.3.3. Needs
xtrabackup-2.3.5 or higher to perform SST (20161028 14:44:28.220)"